Renamed configurations.example to configurations
