--- src/configurations.example.js
@@ -1,2 +1,0 @@
-export const DEFAULT_QUIZ_LINK = "http://localhost:8080/pwa/game/";
-export const DEFAULT_QUIZ_QUES_LINK = "http://localhost:8080/quiz/"
